src/test/java/nextstep/subway/favorite/application/FavoriteService.java
package nextstep.subway.favorite.application;

import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import nextstep.subway.favorite.domain.FavoriteRepository;
import nextstep.subway.favorite.dto.FavoriteRequest;
import nextstep.subway.favorite.dto.FavoriteResponse;

@Service
@Transactional
public class FavoriteService {
	private FavoriteRepository favoriteRepository;

	public FavoriteService(FavoriteRepository favoriteRepository) {
		this.favoriteRepository = favoriteRepository;
	}

	public FavoriteResponse save(FavoriteRequest favoriteRequest) {

		return null;
	}
}
